+++ python/sglang/bench_one_batch_server.py
@@ -14,8 +14,6 @@ import argparse
 import dataclasses
 import itertools
 import json
-import multiprocessing
-import os
 import time
 from typing import Tuple
 
@@ -61,36 +59,6 @@ class BenchArgs:
         return cls(
             **{attr: attr_type(getattr(args, attr)) for attr, attr_type in attrs}
         )
-
-
-def launch_server_process(server_args: ServerArgs):
-    proc = multiprocessing.Process(target=_launch_server_process_internal, args=(server_args,))
-    proc.start()
-    base_url = f"http://{server_args.host}:{server_args.port}"
-    timeout = 600
-
-    start_time = time.time()
-    while time.time() - start_time < timeout:
-        try:
-            headers = {
-                "Content-Type": "application/json; charset=utf-8",
-            }
-            response = requests.get(f"{base_url}/v1/models", headers=headers)
-            if response.status_code == 200:
-                return proc, base_url
-        except requests.RequestException:
-            pass
-        time.sleep(10)
-    raise TimeoutError("Server failed to start within the timeout period.")
-
-
-def _launch_server_process_internal(server_args):
-    try:
-        launch_server(server_args)
-    except Exception as e:
-        raise e
-    finally:
-        kill_process_tree(os.getpid(), include_parent=False)
 
 
 def run_one_case(
